mh_langgraph_workflow/requirements.txt
# LangChain and LangGraph
langchain>=0.3.0
langchain-core>=0.3.0
langgraph>=0.2.0

# API client libraries
fal-client>=0.5.0
requests>=2.31.0
urllib3>=2.0.0

# FastAPI and server
fastapi>=0.115.0
uvicorn>=0.32.0
pydantic>=2.0.0

# Utilities
python-dotenv>=1.0.0
Pillow>=10.0.0

# Optional: for async support
aiohttp>=3.9.0
httpx>=0.27.0
